Rename provider to VaultEconomyProvider
